Add missing forward declaration for get_size_from_typename
The function get_size_from_typename() exists in parseast.c and is used by
codegen.c, but was not declared in cc2.h. This caused cc1 to fail parsing
codegen.c because it couldn't find the symbol in the symbol table.

Added the forward declaration to cc2.h alongside the related function
get_size_from_type_str().

Fixes: cc1 can now successfully parse codegen.c